+++ src/debugging.h
@@ -23,34 +23,41 @@
 
 #pragma once
 
-#include <cassert>
+#include <iomanip>
+#include <iostream>
+#include <sstream>
 
-#include <cmath>
-#include <cstdio>
-#include <ctime>
-#include <string>
-#include "counters.h"
+#include "misc.h"
 #include "types.h"
+
+using namespace std;
 
 namespace ococo {
 
-int print_vcf_header(FILE *file);
+string _pos_stats_uncompr(pos_stats_uncompr_t psu) {
+    stringstream ss;
+    ss << showbase << internal << setfill('0');
+    ss << "[" << nt16_nt256[psu.nt16] << "]"
+       << "(" << psu.counters[0] << "," << psu.counters[1] << ","
+       << psu.counters[2] << "," << psu.counters[3] << ")";
 
-int print_vcf_substitution(FILE *file, const std::string &seq_name, int64_t pos,
-                           char old_base, char new_base,
-                           const pos_stats_uncompr_t &psu);
+    return ss.str();
+}
 
-int print_pileup_line(FILE *file, const std::string &seq_name, int64_t pos,
-                      const pos_stats_uncompr_t &psu);
+template <typename T>
+string _pos_stats_compr(T psc) {
+    stringstream ss;
+    ss << showbase << internal << setfill('0');
+    ss << psc;
 
-void fatal_error(const char *format, ...);
+    return ss.str();
+}
 
-void error(const char *format, ...);
-
-void warning(const char *format, ...);
-
-void info(const char *format, ...);
-
-bool file_exists(const std::string &fn);
+template <typename T>
+void _print_pos_stats(T psc) {
+    pos_stats_uncompr_t psu;
+    psu.decompress(psc);
+    cerr << _pos_stats_compr(psc) << " " << _pos_stats_uncompr(psu) << endl;
+}
 
 }  // namespace ococo
